move themes to separate file and merge then with common
