+++ frontend/package.json
@@ -29,6 +29,7 @@
 		"@babel/runtime": "^7.12.1",
 		"date-fns": "^2.16.1",
 		"history": "^5.0.0",
+		"lodash": "^4.17.20",
 		"polished": "^4.0.3",
 		"react": "^16.14.0",
 		"react-dom": "^16.14.0",
