+++ frontend/src/Theme/index.js
@@ -1,77 +1,5 @@
 import React from 'react';
-
-const lightTheme = {
-	background: 'hsl(0, 0%, 90%)',
-	buttonBg: '#e5e5e5',
-
-	text: {
-		primary: '#000',
-		secondary: '#696969',
-	},
-
-	box: {
-		background: '#FFFFFF',
-		shadow: '2px 2px 0px 2px hsla(0, 0%, 100%, 50%)',
-		highlight: 'rgba(255, 255, 255, 0.5)',
-	},
-
-	topbar: {
-		height: '50px',
-		background: 'hsl(0, 0%, 20%)',
-		text: 'hsl(0, 0%, 70%)',
-		textActive: 'hsl(0, 0%, 90%)',
-	},
-
-	sidebar: {
-		narrow: {
-			width: '50px',
-		},
-
-		wide: {
-			width: '250px',
-		},
-
-		background: '#FFFFFF',
-
-		linkActive: 'hsl(0, 0%, 95%)',
-	},
-};
-
-const darkTheme = {
-	background: 'hsl(0, 0%, 17%)',
-	buttonBg: '#6f6f6f',
-
-	text: {
-		primary: '#EFE9F4',
-		secondary: '#B0B0B0',
-	},
-
-	box: {
-		background: '#505050',
-		shadow: '2px 2px 0px 2px hsla(0, 0%, 34%, 38%)',
-		highlight: 'hsla(0, 0%, 34%, 38%)',
-	},
-
-	topbar: {
-		height: '50px',
-		background: 'hsl(0, 0%, 20%)',
-		text: 'hsl(0, 0%, 70%)',
-		textActive: 'hsl(0, 0%, 85%)',
-	},
-
-	sidebar: {
-		narrow: {
-			width: '50px',
-		},
-
-		wide: {
-			width: '250px',
-		},
-
-		background: '#444444',
-		linkActive: 'hsl(0, 0%, 22%)',
-	},
-};
+import themes from './themes';
 
 // Create react context for theme info
 export const ThemeContext = React.createContext([{}, () => {}, false]);
@@ -101,5 +29,6 @@ export const useTheme = () => {
 	};
 
 	let isDark = theme === 'dark';
-	return [isDark ? darkTheme : lightTheme, switchTheme, isDark];
+
+	return [isDark ? themes.dark : themes.light, switchTheme, isDark];
 };
